add device into into view profile item, add to attach
